Create a MainSideDrawer component
Move from App.js
